\name{gWidgets-undocumented}
\alias{gWidgets-undocumented.Rd}
\alias{show,guiWidget-method}
\alias{editFormulaDialog}
\alias{editSelectDialog}
\alias{editSubsetDialog}
\alias{.fixFontMessUp} %% screwed this up.
\alias{str2}
\alias{gwCat}			% for error messages

%% These from the ANY class for compound widgets
\alias{.add,ANY,ANY,gWidgetANY-method}
\alias{.add,gHelpANY,ANY,character-method}
\alias{.add,gHelpANY,ANY,list-method}
\alias{.add,gHelpANY-method}
\alias{add,ANY-method}
\alias{[<-,gWidgetANY-method}
\alias{[,gWidgetANY-method}
\alias{[,gCommandlineANY-method}
\alias{dimnames<-,gWidgetANY-method}
\alias{dimnames,gWidgetANY-method}
\alias{.dispose,gHelpANY-method}
\alias{enabled<-,gWidgetANY-method}
\alias{enabled,gWidgetANY-method}
\alias{focus<-,gWidgetANY-method}
\alias{focus,gWidgetANY-method}
\alias{font<-,gWidgetANY-method}
\alias{.gcommandline,ANY-method}
\alias{.ggenericwidget,ANY-method}
\alias{.ghelpbrowser,ANY-method}
\alias{.ghelp,ANY-method}
\alias{.leftBracket,gCommandlineANY-method}
\alias{.length,gHelpANY-method}
\alias{names<-,gWidgetANY-method}
\alias{names,gWidgetANY-method}
\alias{.size,gHelpbrowserANY-method}
\alias{size<-,gWidgetANY-method}
\alias{.size<-,gHelpbrowserANY,ANY,numeric-method}
\alias{.svalue<-,gCommandlineANY-method}
\alias{.svalue<-,gAddargANY-method}
\alias{.svalue<-,gBivariateANY-method}
\alias{.svalue<-,gModelANY-method}
\alias{.svalue,character-method}
\alias{.svalue,gCommandlineANY-method}
\alias{.svalue,gGenericWidgetANY-method}
\alias{.svalue,gHelpANY-method}
\alias{.svalue,gAddargANY-method}
\alias{.svalue,gUnivariateANY-method}
\alias{.svalue,gUnivariateTableANY-method}
\alias{.svalue,gFileURLANY-method}
\alias{.svalue,gBivariateANY-method}
\alias{.svalue,gModelANY-method}
\alias{.svalue,gLmerANY-method}
\alias{svalue<-,gWidgetANY-method}
\alias{svalue,numeric-method}
\alias{svalue,character-method}
\alias{svalue,gWidgetANY-method}
\alias{.tag<-,guiWidget-method}
\alias{.tag<-,gWidgetANY-method}
\alias{.tag,guiWidget-method}
\alias{.tag,gWidgetANY-method}
\alias{tag<-,gWidgetANY-method}
\alias{tag,gWidgetANY-method}
\alias{update,gWidgetANY-method}
\alias{visible<-,gWidgetANY-method}
\alias{visible,gWidgetANY-method}
\alias{.visible<-,gHelpbrowserANY,ANY,logical-method}
\alias{isExtant,gWidgetANY-method}

\alias{.gsvg,ANY-method}
\alias{.redo,guiWidget-method}
\alias{.redo,gWidgetANY-method}
\alias{.svalue,NULL-method}
\alias{.toolkitProvidesWidget,ANY-method}
\alias{.undo,guiWidget-method}
\alias{.undo,gWidgetANY-method}

\alias{.gformlayout,ANY-method}
\alias{.leftBracket,gFormLayoutANY-method}
\alias{.names,gFormLayoutANY-method}
\alias{.svalue,gFormLayoutANY-method}
\alias{[,gFormLayoutANY-method}




%% these from ANY methods
\title{Undocumented, but exported, functions}
\description{Undocumented, but exported functions from gWidgets.}
\keyword{interface}% at least one, from doc/KEYWORDS
\keyword{internal}